+++ projects/04-llm-adapter-shadow/src/llm_adapter/runner_async_support/__init__.py
@@ -1,0 +1,10 @@
+"""Support package for :mod:`llm_adapter.runner_async`."""
+from .failures import emit_consensus_failure
+from .invoker import AsyncProviderInvoker
+from .shadow_logging import build_shadow_log_metadata
+
+__all__ = [
+    "AsyncProviderInvoker",
+    "build_shadow_log_metadata",
+    "emit_consensus_failure",
+]
